+++ e2e/kit/houdini.config.js
@@ -27,6 +27,9 @@ const config = {
   types: {
     RentedBook: {
       keys: ['userId', 'bookId']
+    },
+    UnionAorB: {
+      keys: []
     }
   },