+++ packages/houdini/src/codegen/generators/typescript/typescript.test.ts
@@ -946,8 +946,8 @@ describe('typescript', function () {
 			                            };
 			                        };
 			                        "typeMap": {
-			                            "Cat": "Node";
 			                            "User": "Node";
+			                            "Cat": "Node";
 			                        };
 			                    };
 			                    "fields": {
@@ -2052,6 +2052,11 @@ describe('typescript', function () {
 			                                };
 			                            };
 			                            "Cat": {
+			                                "isAnimal": {
+			                                    "type": "Boolean";
+			                                    "keyRaw": "isAnimal";
+			                                    "visible": true;
+			                                };
 			                                "kitty": {
 			                                    "type": "Boolean";
 			                                    "keyRaw": "kitty";
@@ -2060,11 +2065,6 @@ describe('typescript', function () {
 			                                "id": {
 			                                    "type": "ID";
 			                                    "keyRaw": "id";
-			                                    "visible": true;
-			                                };
-			                                "isAnimal": {
-			                                    "type": "Boolean";
-			                                    "keyRaw": "isAnimal";
 			                                    "visible": true;
 			                                };
 			                                "__typename": {